MNT: allow use of FakePytmcSignal outside of tests
